Created motorRotator.h to call/import  from main
Goal is to have all of our modules be called from main.ino.
Todo: test if importing source code from a .ino file is allowed, or if we need to remove the setup() and loop() functions from motorRotator.ino now that we import it.
(maybe not since the .h file doesn't specify those as imports?)