Loop thru angle ranges to find strength to apply
Todo:
Change areLengthsMatching to work with the current angle_ranges_arr
(it's not an array of pairs,
so the length won't
directly match the strengths_arr)

Maybe it will just be 1 element longer?
Then we could use a version of this
to check matching lens.
`len(angle_ranges_arr) == len(strengths_arr) + 1`