Attempt larger angle range. Todo: maybe Y angle
 w/ less subdivisions (i.e. 2 angles, 1 strength)

Liz recommended these angles, so I'm committing this here so we remember we may need to detect
quadrant 3, -135 to -90
(or 235 to 270
(if the x axis is aways angle 0) or -45 to -90)

^ that last scenario is the real one because
90 degrees reads as xAngle = -90
180 degrees reads as xAngle = 0
270 degrees reads as xAngle = -90
360 degrees reads as xAngle = 0

We want to focus on the changes in quadrant three where the swing is moving "right to left"
so going from 0 toward -90.

The problem is that the same condition is true when the swing is in quadrant 4 moving "left to right".

**Todo**
- incorporate yAngle or (more likely) zAngle to detect that we are changing quadrants.