Delete interpretAcc since it's part of secondary.
This used to be completely separate as we were trying to think of ways to interpret the data from the accelerometer.

If we need it again we can go to back to this commit: https://github.com/yfortaleza1/ElectroSwing/commit/626056863adfb709cef62f1339edd47b8ea519b1

Or we can just look at the interpret-accelerometer branch.