Update secondary.ino
Need to get motor working with interrupts? (is that even necessary, since it works fine if coded in main. if it was done in main, then no trig and stuff would be done because so it would be in the motorMove code basically with whatever delay is going on)
Determining which motor movement was started to be coded but needs to be finished
Check accelerometer Test functionality with LEDs
Everything that's angle based needs to be tested with full swing setup